target/requirements.txt
alembic==1.0.0
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flup==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
mysqlclient==1.3.13
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
validate-email==1.3
Werkzeug==0.15.2
